backend/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
whitenoise==6.9.0
